move listPropsAndValues and generateTooltip into utils.js
